+++ IsekaiMod/Content/Classes/IsekaiProtagonist/IsekaiProtagonistClass.cs
@@ -14,12 +14,12 @@ using UnityEngine;
 using static IsekaiMod.Main;
 
 namespace IsekaiMod.Content.Classes.IsekaiProtagonist {
+
     internal class IsekaiProtagonistClass {
         private static BlueprintCharacterClass isekaiProtagonistClass;
 
         // Icons
-        private static readonly Sprite Icon_SneakAttack = BlueprintTools.GetBlueprint<BlueprintFeature>("9b9eac6709e1c084cb18c3a366e0ec87").m_Icon;
-
+        private static readonly Sprite Icon_AllSkilled = BlueprintTools.GetBlueprint<BlueprintFeature>("f3bc6f9c855b2fb4e9aea364b8163aca").m_Icon;
         private static readonly Sprite Icon_ForetellAidBuff = BlueprintTools.GetBlueprint<BlueprintBuff>("faf473e3a977fd4428cd3f1a526346d2").m_Icon;
         private static readonly Sprite Icon_EdictOfImpenetrableFortress = BlueprintTools.GetBlueprint<BlueprintAbility>("d7741c08ccf699e4a8a8f8ab2ed345f8").m_Icon;
         private static readonly Sprite Icon_TrickFate = BlueprintTools.GetBlueprint<BlueprintAbility>("6e109d21da9e1c44fb772a9eca2cafdd").m_Icon;
@@ -27,12 +27,10 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist {
 
         // Stat Progression
         private static readonly BlueprintStatProgression BABFull = BlueprintTools.GetBlueprint<BlueprintStatProgression>("b3057560ffff3514299e8b93e7648a9d");
-
         private static readonly BlueprintStatProgression SavesHigh = BlueprintTools.GetBlueprint<BlueprintStatProgression>("ff4662bde9e75f145853417313842751");
 
         // Used in Class
         private static readonly BlueprintCharacterClass SlayerClass = BlueprintTools.GetBlueprint<BlueprintCharacterClass>("c75e0971973957d4dbad24bc7957e4fb");
-
         private static readonly BlueprintCharacterClass AnimalClass = BlueprintTools.GetBlueprint<BlueprintCharacterClass>("4cd1757a0eea7694ba5c933729a53920");
 
         public static void Add() {
@@ -55,13 +53,20 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist {
             var IsekaiProtagonistOverpoweredAbilityFeat = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "IsekaiProtagonistOverpoweredAbilityFeat", bp => {
                 bp.SetName(IsekaiContext, "Overpowered Abilities");
                 bp.SetDescription(IsekaiContext, "Isekai Protagonists gain Overpowered Abilities as they increase their level. These are extremely powerful abiilities that surpass even gods.");
-                bp.m_DescriptionShort = bp.m_Description; 
+                bp.m_DescriptionShort = bp.m_Description;
                 bp.m_Icon = Icon_TrickFate;
             });
             var IsekaiProtagonistBonusFeat = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "IsekaiProtagonistBonusFeat", bp => {
                 bp.SetName(IsekaiContext, "Bonus Feat");
                 bp.SetDescription(IsekaiContext, "Isekai Protagonists gain twice as many {g|Encyclopedia:Feat}feats{/g} as the other classes.");
-                bp.m_DescriptionShort = bp.m_Description; bp.m_Icon = Icon_BasicFeatSelection;
+                bp.m_DescriptionShort = bp.m_Description;
+                bp.m_Icon = Icon_BasicFeatSelection;
+            });
+            var IsekaiProtagonistLegacyFeat = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "IsekaiProtagonistLegacyFeat", bp => {
+                bp.SetName(IsekaiContext, "Legacy");
+                bp.SetDescription(IsekaiContext, "Isekai Protagonists gain a limited set of features from other classes.");
+                bp.m_DescriptionShort = bp.m_Description;
+                bp.m_Icon = Icon_AllSkilled;
             });
 
             // Main Class
@@ -105,11 +110,12 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist {
                 bp.SecondaryColor = 9;
                 bp.MaleEquipmentEntities = SlayerClass.MaleEquipmentEntities;
                 bp.FemaleEquipmentEntities = SlayerClass.FemaleEquipmentEntities;
-                bp.m_SignatureAbilities = new BlueprintFeatureReference[4] {
+                bp.m_SignatureAbilities = new BlueprintFeatureReference[5] {
                     IsekaiProtagonistPlotArmorFeat.ToReference<BlueprintFeatureReference>(),
                     IsekaiProtagonistSpecialPowerFeat.ToReference<BlueprintFeatureReference>(),
                     IsekaiProtagonistOverpoweredAbilityFeat.ToReference<BlueprintFeatureReference>(),
                     IsekaiProtagonistBonusFeat.ToReference<BlueprintFeatureReference>(),
+                    IsekaiProtagonistLegacyFeat.ToReference<BlueprintFeatureReference>(),
                 };
                 bp.AddComponent<PrerequisiteNoClassLevel>(c => {
                     c.m_CharacterClass = AnimalClass.ToReference<BlueprintCharacterClassReference>();
@@ -118,8 +124,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist {
                     c.Not = true;
                     c.HideInUI = true;
                 });
-                if (IsekaiContext.AddedContent.ExcludeCompanionsFromIsekaiClass)
-                {
+                if (IsekaiContext.AddedContent.ExcludeCompanionsFromIsekaiClass) {
                     bp.AddComponent<PrerequisiteIsMainCharacter>();
                 }
 
@@ -137,24 +142,29 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist {
             // Register Class
             ThingsNotHandledByTTTCore.RegisterClass(isekaiProtagonistClass);
         }
+
         public static void RegisterArchetype(BlueprintArchetype archetype) {
             BlueprintCharacterClass IsekaiProtagonistClass = Get();
             IsekaiProtagonistClass.m_Archetypes = IsekaiProtagonistClass.m_Archetypes.AppendToArray(archetype.ToReference<BlueprintArchetypeReference>());
         }
+
         public static void SetProgression(BlueprintProgression progression) {
             BlueprintCharacterClass IsekaiProtagonistClass = Get();
             IsekaiProtagonistClass.m_Progression = progression.ToReference<BlueprintProgressionReference>();
         }
+
         public static void SetDefaultBuild(BlueprintUnitFact prebuildFeatureList) {
             BlueprintCharacterClass IsekaiProtagonistClass = Get();
             IsekaiProtagonistClass.m_DefaultBuild = prebuildFeatureList.ToReference<BlueprintUnitFactReference>();
         }
+
         public static BlueprintCharacterClass Get() {
-            if (isekaiProtagonistClass != null) { 
-                return isekaiProtagonistClass; 
+            if (isekaiProtagonistClass != null) {
+                return isekaiProtagonistClass;
             }
             return BlueprintTools.GetModBlueprint<BlueprintCharacterClass>(IsekaiContext, "IsekaiProtagonistClass");
         }
+
         public static BlueprintCharacterClassReference GetReference() {
             return Get().ToReference<BlueprintCharacterClassReference>();
         }