+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added `Areshkagal` to `Summon Calamity` ability.
+### Fixed
+- Fixed Deathsnatcher default name.
 
 ## [4.0.0] - 2023-03-12
 ### Added