+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Areshkagal` to `Summon Calamity` ability.
 ### Fixed
 - Fixed Deathsnatcher default name.
+- Fixed `Supreme Being` to allow base stat increases.
 
 ## [4.0.0] - 2023-03-12
 ### Added