+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed Deathsnatcher default name.
 - Fixed `Supreme Being` to allow base stat increases.
+- Fixed `Instakill` Spell penetration modifier.
 
 ## [4.0.0] - 2023-03-12
 ### Added