+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added `Isekai Furry` heritage (kitsune heritage).
 - Added `Areshkagal` to `Summon Calamity` ability.
 - Added `Legacy` as a listed Isekai Protagonist signature class feature.
 - Added icons for `Legacy` and `Dual Legacy` selections.