+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [4.1.0] - 2023-03-15
 ### Added
 - Added `Isekai Furry` heritage (kitsune heritage).
 - Added `Areshkagal` to `Summon Calamity` ability.
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added icons for `Legacy` and `Dual Legacy` selections.
 ### Changed
 - Buffed `Killing Intent` Special power fear duration from 1 round to 1d4 rounds.
+- Changed `Dual Legacy` so that it can only be selected once.
 - Renamed `Legacy Class Feature` to `Legacy`.
 - Renamed `Dual Class - Legacy Class` Feature to `Dual Legacy`.
 ### Fixed